# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: The Nuptial Plan – Wedding planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /weddings/{weddingId}/guests:
    get:
      operationId: listGuests
      tags: [guests]
      summary: List guests for a wedding
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of guests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guest"
    post:
      operationId: createGuest
      tags: [guests]
      summary: Add a guest to a wedding
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
  /weddings/{weddingId}/guests/{id}:
    patch:
      operationId: updateGuest
      tags: [guests]
      summary: Update a guest
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
    delete:
      operationId: deleteGuest
      tags: [guests]
      summary: Delete a guest
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /weddings/{weddingId}/guests/stats:
    get:
      operationId: getGuestStats
      tags: [guests]
      summary: RSVP stats for a wedding
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Guest stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestStats"

  # ─── Budget ───────────────────────────────────────────────────
...
  /weddings/{weddingId}/events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List calendar events for a wedding
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a calendar event
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarEventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
  /weddings/{weddingId}/events/{id}:
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update a calendar event
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarEventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete a calendar event
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── Contracts ────────────────────────────────────────────────
  /weddings/{weddingId}/contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List contracts for a wedding
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Create a contract
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /weddings/{weddingId}/contracts/{id}:
    patch:
      operationId: updateContract
      tags: [contracts]
      summary: Update a contract
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
    delete:
      operationId: deleteContract
      tags: [contracts]
      summary: Delete a contract
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── Payments ─────────────────────────────────────────────────
  /weddings/{weddingId}/payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments for a wedding
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /weddings/{weddingId}/payments/{id}:
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update a payment
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    delete:
      operationId: deletePayment
      tags: [payments]
      summary: Delete a payment
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: number
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Guest:
      type: object
      required: [id, weddingId, name, rsvpStatus, createdAt]
      properties:
        id:
          type: number
        weddingId:
          type: number
        name:
          type: string
        email:
          type: ["string", "null"]
        tableNumber:
          type: ["string", "null"]
        dietaryRequirements:
          type: ["string", "null"]
        rsvpStatus:
          type: string
          enum: [confirmed, pending, declined]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    GuestInput:
      type: object
      required: [name, rsvpStatus]
      properties:
        name:
          type: string
        email:
          type: string
        tableNumber:
          type: string
        dietaryRequirements:
          type: string
        rsvpStatus:
          type: string
          enum: [confirmed, pending, declined]
        notes:
          type: string
    GuestUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        tableNumber:
          type: string
        dietaryRequirements:
          type: string
        rsvpStatus:
          type: string
          enum: [confirmed, pending, declined]
        notes:
          type: string
    GuestStats:
      type: object
      required: [total, confirmed, pending, declined]
      properties:
        total:
          type: number
        confirmed:
          type: number
        pending:
          type: number
        declined:
          type: number

    # ─── Budget ───────────────────────────────────────────────
...
    CalendarEvent:
      type: object
      required: [id, weddingId, title, eventDate, createdAt]
      properties:
        id:
          type: number
        weddingId:
          type: number
        title:
          type: string
        detail:
          type: ["string", "null"]
        eventDate:
          type: string
          description: ISO date string (YYYY-MM-DD)
        eventTime:
          type: ["string", "null"]
          description: HH:MM format
        tone:
          type: ["string", "null"]
          enum: [gold, rose, sage, null]
        completed:
          type: boolean
        createdAt:
          type: string
    CalendarEventInput:
      type: object
      required: [title, eventDate]
      properties:
        title:
          type: string
        detail:
          type: string
        eventDate:
          type: string
        eventTime:
          type: string
        tone:
          type: string
          enum: [gold, rose, sage]
        completed:
          type: boolean
    CalendarEventUpdate:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        eventDate:
          type: string
        eventTime:
          type: string
        tone:
          type: string
          enum: [gold, rose, sage]
        completed:
          type: boolean

    # ─── Contract ─────────────────────────────────────────────
    Contract:
      type: object
      required: [id, weddingId, vendorName, status, totalAmountCents, createdAt]
      properties:
        id:
          type: number
        weddingId:
          type: number
        vendorId:
          type: ["number", "null"]
        vendorName:
          type: string
        status:
          type: string
          enum: [signed, pending, partial, cancelled]
        totalAmountCents:
          type: number
        depositPaidCents:
          type: ["number", "null"]
        signedDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ContractInput:
      type: object
      required: [vendorName, status, totalAmountCents]
      properties:
        vendorId:
          type: number
        vendorName:
          type: string
        status:
          type: string
          enum: [signed, pending, partial, cancelled]
        totalAmountCents:
          type: number
        depositPaidCents:
          type: number
        signedDate:
          type: string
        notes:
          type: string
    ContractUpdate:
      type: object
      properties:
        vendorName:
          type: string
        status:
          type: string
          enum: [signed, pending, partial, cancelled]
        totalAmountCents:
          type: number
        depositPaidCents:
          type: number
        signedDate:
          type: string
        notes:
          type: string

    # ─── Payment ──────────────────────────────────────────────
    Payment:
      type: object
      required: [id, weddingId, vendorName, description, amountCents, dueDate, status, createdAt]
      properties:
        id:
          type: number
        weddingId:
          type: number
        vendorId:
          type: ["number", "null"]
        vendorName:
          type: string
        description:
          type: string
        amountCents:
          type: number
        dueDate:
          type: string
          description: ISO date string (YYYY-MM-DD)
        status:
          type: string
          enum: [pending, paid, overdue, scheduled]
        paidDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [vendorName, description, amountCents, dueDate, status]
      properties:
        vendorId:
          type: number
        vendorName:
          type: string
        description:
          type: string
        amountCents:
          type: number
        dueDate:
          type: string
        status:
          type: string
          enum: [pending, paid, overdue, scheduled]
        paidDate:
          type: string
        notes:
          type: string
    PaymentUpdate:
      type: object
      properties:
        vendorName:
          type: string
        description:
          type: string
        amountCents:
          type: number
        dueDate:
          type: string
        status:
          type: string
          enum: [pending, paid, overdue, scheduled]
        paidDate:
          type: string
        notes:
          type: string