# lib/api-zod/src/generated/types/contractUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * The Nuptial Plan – Wedding planner API
 * OpenAPI spec version: 0.1.0
 */
import type { ContractUpdateStatus } from './contractUpdateStatus';

export interface ContractUpdate {
  vendorName?: string;
  status?: ContractUpdateStatus;
  totalAmountCents?: number;
  depositPaidCents?: number;
  signedDate?: string;
  notes?: string;
}