# lib/api-zod/src/generated/types/guestUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * The Nuptial Plan – Wedding planner API
 * OpenAPI spec version: 0.1.0
 */
import type { GuestUpdateRsvpStatus } from './guestUpdateRsvpStatus';

export interface GuestUpdate {
  name?: string;
  email?: string;
  tableNumber?: string;
  dietaryRequirements?: string;
  rsvpStatus?: GuestUpdateRsvpStatus;
  notes?: string;
}